Passes 5 tests

+++ Projects/Raycast_Kinematic_Function/Raycast_Kinematic_Function.cpp
@@ -45,15 +45,13 @@ TrajectoryResult PredictTrajectory(const Vec3& start_position,
         Physics::QueryResult raycast_result = Physics::Raycast(current_position, current_position + current_velocity * raycast_time_step);
 
         // Calculate the time it takes to reach the hit position between two consecutive raycasts
-        double time_to_hit = raycast_time_step;
+        float time_to_hit = raycast_time_step;
 
         if (raycast_result.m_ValidHit) {
             valid_hit = true;
-            // Calculate the time adjustment based on the actual hit position and the distance traveled
             double hit_distance = magnitude(raycast_result.m_HitPos - current_position);
             double total_distance = magnitude(current_velocity) * raycast_time_step;
             time_to_hit *= hit_distance / total_distance;
-
             result.m_EndPoint = raycast_result.m_HitPos;
             result.m_Time = current_time + time_to_hit;
             break;
@@ -71,6 +69,7 @@ TrajectoryResult PredictTrajectory(const Vec3& start_position,
         current_velocity.x += up_vector.x * gravity_accel * time_to_hit;
         current_velocity.y += up_vector.y * gravity_accel * time_to_hit;
         current_velocity.z += up_vector.z * gravity_accel * time_to_hit;
+        
     }
 
     if (!valid_hit) {
